+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/MPGD/trapezoid/MPGDTrapezoidGeant4Factory.java
@@ -1,4 +1,4 @@
-package org.jlab.detector.geant4.v2.MPGD2.trapezoid;
+package org.jlab.detector.geant4.v2.MPGD.trapezoid;
 
 import eu.mihosoft.vrl.v3d.Vector3d;
 import org.jlab.detector.geant4.v2.Geant4Factory;
